+++ packages/google_mobile_ads/lib/src/ad_instance_manager.dart
@@ -718,16 +718,35 @@ class AdMessageCodec extends StandardMessageCodec {
   static const int _valueVideoOptions = 145;
   static const int _valueRequestConfigurationParams = 146;
   static const int _valueInlineAdaptiveBannerAdSize = 147;
+  static const int _valueLocationParams = 148;
 
   @override
   void writeValue(WriteBuffer buffer, dynamic value) {
     if (value is AdSize) {
       writeAdSize(buffer, value);
+    } else if (value is AdManagerAdRequest) {
+      buffer.putUint8(_valueAdManagerAdRequest);
+      writeValue(buffer, value.keywords);
+      writeValue(buffer, value.contentUrl);
+      writeValue(buffer, value.customTargeting);
+      writeValue(buffer, value.customTargetingLists);
+      writeValue(buffer, value.nonPersonalizedAds);
+      writeValue(buffer, value.neighboringContentUrls);
+      if (defaultTargetPlatform == TargetPlatform.android) {
+        writeValue(buffer, value.httpTimeoutMillis);
+      }
+      writeValue(buffer, value.publisherProvidedId);
+      writeValue(buffer, value.location);
     } else if (value is AdRequest) {
       buffer.putUint8(_valueAdRequest);
       writeValue(buffer, value.keywords);
       writeValue(buffer, value.contentUrl);
       writeValue(buffer, value.nonPersonalizedAds);
+      writeValue(buffer, value.neighboringContentUrls);
+      if (defaultTargetPlatform == TargetPlatform.android) {
+        writeValue(buffer, value.httpTimeoutMillis);
+      }
+      writeValue(buffer, value.location);
     } else if (value is RewardItem) {
       buffer.putUint8(_valueRewardItem);
       writeValue(buffer, value.amount);
@@ -755,13 +774,6 @@ class AdMessageCodec extends StandardMessageCodec {
       writeValue(buffer, value.code);
       writeValue(buffer, value.domain);
       writeValue(buffer, value.message);
-    } else if (value is AdManagerAdRequest) {
-      buffer.putUint8(_valueAdManagerAdRequest);
-      writeValue(buffer, value.keywords);
-      writeValue(buffer, value.contentUrl);
-      writeValue(buffer, value.customTargeting);
-      writeValue(buffer, value.customTargetingLists);
-      writeValue(buffer, value.nonPersonalizedAds);
     } else if (value is AdapterInitializationState) {
       buffer.putUint8(_valueInitializationState);
       writeValue(buffer, describeEnum(value));
@@ -796,6 +808,14 @@ class AdMessageCodec extends StandardMessageCodec {
       writeValue(buffer, value.tagForChildDirectedTreatment);
       writeValue(buffer, value.tagForUnderAgeOfConsent);
       writeValue(buffer, value.testDeviceIds);
+    } else if (value is LocationParams) {
+      buffer.putUint8(_valueLocationParams);
+      writeValue(buffer, value.accuracy);
+      writeValue(buffer, value.longitude);
+      writeValue(buffer, value.latitude);
+      if (defaultTargetPlatform == TargetPlatform.android) {
+        writeValue(buffer, value.time);
+      }
     } else {
       super.writeValue(buffer, value);
     }
@@ -855,6 +875,12 @@ class AdMessageCodec extends StandardMessageCodec {
           keywords: readValueOfType(buffer.getUint8(), buffer)?.cast<String>(),
           contentUrl: readValueOfType(buffer.getUint8(), buffer),
           nonPersonalizedAds: readValueOfType(buffer.getUint8(), buffer),
+          neighboringContentUrls:
+              readValueOfType(buffer.getUint8(), buffer)?.cast<String>(),
+          httpTimeoutMillis: (defaultTargetPlatform == TargetPlatform.android)
+              ? readValueOfType(buffer.getUint8(), buffer)
+              : null,
+          location: readValueOfType(buffer.getUint8(), buffer),
         );
       case _valueRewardItem:
         return RewardItem(
@@ -897,6 +923,13 @@ class AdMessageCodec extends StandardMessageCodec {
             readValueOfType(buffer.getUint8(), buffer),
           ),
           nonPersonalizedAds: readValueOfType(buffer.getUint8(), buffer),
+          neighboringContentUrls:
+              readValueOfType(buffer.getUint8(), buffer)?.cast<String>(),
+          httpTimeoutMillis: (defaultTargetPlatform == TargetPlatform.android)
+              ? readValueOfType(buffer.getUint8(), buffer)
+              : null,
+          publisherProvidedId: readValueOfType(buffer.getUint8(), buffer),
+          location: readValueOfType(buffer.getUint8(), buffer),
         );
       case _valueInitializationState:
         switch (readValueOfType(buffer.getUint8(), buffer)) {
@@ -955,6 +988,15 @@ class AdMessageCodec extends StandardMessageCodec {
           tagForUnderAgeOfConsent: readValueOfType(buffer.getUint8(), buffer),
           testDeviceIds:
               readValueOfType(buffer.getUint8(), buffer).cast<String>(),
+        );
+      case _valueLocationParams:
+        return LocationParams(
+          accuracy: readValueOfType(buffer.getUint8(), buffer),
+          longitude: readValueOfType(buffer.getUint8(), buffer),
+          latitude: readValueOfType(buffer.getUint8(), buffer),
+          time: (defaultTargetPlatform == TargetPlatform.android)
+              ? readValueOfType(buffer.getUint8(), buffer)
+              : null,
         );
       default:
         return super.readValueOfType(type, buffer);